Establish class hierarchy for property usages
